Refactor bucket classes
- add api docs for public methods and fields
- separate bucket implementation into their own files
- rename some methods into their non-abbreviated versions
