+++ src/main/scala/com/exasol/cloudetl/bucket/AzureBlobBucket.scala
@@ -1,0 +1,40 @@
+package com.exasol.cloudetl.bucket
+
+import org.apache.hadoop.conf.Configuration
+
+/** A [[Bucket]] implementation for the Azure Blob Storage */
+final case class AzureBlobBucket(path: String, params: Map[String, String]) extends Bucket {
+
+  /** @inheritdoc */
+  override val bucketPath: String = path
+
+  /** @inheritdoc */
+  override def validate(): Unit =
+    Bucket.validate(params, Bucket.AZURE_PARAMETERS)
+
+  /**
+   * @inheritdoc
+   *
+   * Additionally validates that all required parameters are available
+   * in order to create a configuration.
+   */
+  override def createConfiguration(): Configuration = {
+    validate()
+
+    val conf = new Configuration()
+    val accountName = Bucket.requiredParam(params, "AZURE_ACCOUNT_NAME")
+    val accountSecretKey = Bucket.requiredParam(params, "AZURE_SECRET_KEY")
+    conf.set("fs.azure", classOf[org.apache.hadoop.fs.azure.NativeAzureFileSystem].getName)
+    conf.set("fs.wasb.impl", classOf[org.apache.hadoop.fs.azure.NativeAzureFileSystem].getName)
+    conf.set("fs.wasbs.impl", classOf[org.apache.hadoop.fs.azure.NativeAzureFileSystem].getName)
+    conf.set("fs.AbstractFileSystem.wasb.impl", classOf[org.apache.hadoop.fs.azure.Wasb].getName)
+    conf.set(
+      "fs.AbstractFileSystem.wasbs.impl",
+      classOf[org.apache.hadoop.fs.azure.Wasbs].getName
+    )
+    conf.set(s"fs.azure.account.key.$accountName.blob.core.windows.net", accountSecretKey)
+
+    conf
+  }
+
+}
