+++ src/main/scala/com/exasol/cloudetl/bucket/GCSBucket.scala
@@ -1,0 +1,36 @@
+package com.exasol.cloudetl.bucket
+
+import org.apache.hadoop.conf.Configuration
+
+/** A [[Bucket]] implementation for the Google Cloud Storage (GCS) */
+final case class GCSBucket(path: String, params: Map[String, String]) extends Bucket {
+
+  /** @inheritdoc */
+  override val bucketPath: String = path
+
+  /** @inheritdoc */
+  override def validate(): Unit =
+    Bucket.validate(params, Bucket.GCS_PARAMETERS)
+
+  /**
+   * @inheritdoc
+   *
+   * Additionally validates that all required parameters are available
+   * in order to create a configuration.
+   */
+  override def createConfiguration(): Configuration = {
+    validate()
+
+    val conf = new Configuration()
+    conf.set("fs.gs.impl", classOf[com.google.cloud.hadoop.fs.gcs.GoogleHadoopFileSystem].getName)
+    conf.setBoolean("fs.gs.auth.service.account.enable", true)
+    conf.set("fs.gs.project.id", Bucket.requiredParam(params, "GCS_PROJECT_ID"))
+    conf.set(
+      "fs.gs.auth.service.account.json.keyfile",
+      Bucket.requiredParam(params, "GCS_KEYFILE_PATH")
+    )
+
+    conf
+  }
+
+}
